Put Drawers into Dressers
Dressers allow drawers to float next to each other.
Ghost drawers didn't seem to work as expected (will remove).
Adding a pseudo-element with a calculated width
(to act as a ghost for the handle)
works on the right dresser, but not on the left dresser.
While floating offers the ideal position to elements
that are ordered first among its siblings,
tranformations on elements not only remove the element
from the normal document flow, but also offers
the highest stacking order to elements that are
ordered last among its siblings.